Create random password generator
I have executed python task. In this I have prompted the user to specify the desired length to the password. I have use a strong combination of random characters to
generate a password of the specified length and then finally printed the generated password on the screen
Thank You.
I hope that your secrets will be safe with help of my password.